+++ lib/schedulers.dart
@@ -1,7 +1,8 @@
-// SPDX-FileCopyrightText: (c) 2021 Artёm IG <github.com/rtmigo>
+// SPDX-FileCopyrightText: (c) 2021 Artsiom iG <github.com/rtmigo>
 // SPDX-License-Identifier: MIT
 
-export 'src/50_interval.dart';
-export 'src/50_rateLimiting.dart';
-export 'src/50_lazy.dart';
-export 'src/50_time.dart';
+export 'src/c_interval.dart';
+export 'src/c_rate_limiting.dart';
+export 'src/c_lazy.dart';
+export 'src/c_time.dart';
+export 'src/c_concurrent.dart';