+++ test/lazy_test.dart
@@ -1,18 +1,20 @@
-// SPDX-FileCopyrightText: (c) 2021 Artёm IG <github.com/rtmigo>
+// SPDX-FileCopyrightText: (c) 2021 Artsiom iG <github.com/rtmigo>
 // SPDX-License-Identifier: MIT
 
 import 'package:schedulers/schedulers.dart';
 import 'package:test/test.dart';
 
 void main() {
+  // todo test tasks that throw exceptions
+
   test('lazy 1', () async {
 
     final scheduler = LazyScheduler(latency: const Duration(milliseconds: 100));
 
     int x=0;
     int y=0;
-    final funcA = ()=>x++;
-    final funcB = ()=>y++;
+    int funcA()=>x++;
+    int funcB()=>y++;
 
     for (int i=0; i<100; ++i) {
       scheduler.run(funcA);
@@ -20,9 +22,9 @@ void main() {
 
     expect(x, 0);
 
-    await Future.delayed(const Duration(milliseconds: 10));
+    await Future<void>.delayed(const Duration(milliseconds: 10));
     expect(x, 0);
-    await Future.delayed(const Duration(milliseconds: 100));
+    await Future<void>.delayed(const Duration(milliseconds: 100));
     expect(x, 1);
 
     for (int i=0; i<100; ++i) {
@@ -32,11 +34,11 @@ void main() {
     expect(x, 1);
     expect(y, 0);
 
-    await Future.delayed(const Duration(milliseconds: 100));
+    await Future<void>.delayed(const Duration(milliseconds: 100));
     expect(x, 1);
     expect(y, 1);
 
-    await Future.delayed(const Duration(milliseconds: 100));
+    await Future<void>.delayed(const Duration(milliseconds: 100));
     expect(x, 1);
     expect(y, 1);
   });